TDD/study/java-blackjack/src/main/java/dev/hongsii/blackjack/model/Hands.java
package dev.hongsii.blackjack.model;

import lombok.AccessLevel;
import lombok.AllArgsConstructor;

import java.util.ArrayList;
import java.util.Collections;
import java.util.List;

@AllArgsConstructor(access = AccessLevel.PRIVATE)
public class Hands {

    private List<Card> cards;

    public static Hands initialize() {
        return of(new ArrayList<>());
    }

    public static Hands of(List<Card> cards) {
        return new Hands(cards);
    }

    public int getTotalScore() {
        return cards.stream()
                .mapToInt(Card::getScore)
                .sum();
    }

    public void addCard(Card card) {
        cards.add(card);
    }

    public List<Card> getCards() {
        return Collections.unmodifiableList(cards);
    }
}
